services/image-checker/requirements.txt
fastapi
uvicorn[standard]
Pillow
imagehash
numpy
requests
python-multipart
